tabfm: R Interface to 'TabFM', a Zero-Shot Foundation Model for Tabular Data

A thin wrapper around the 'TabFM' 'Python' package from Google Research (https://github.com/google-research/tabfm). Provides zero-shot classification and regression on tabular data via in-context learning, with no feature engineering or hyperparameter tuning.

See also

Author

Maintainer: Matt O'Reilly fermoymatt@gmail.com

Authors:

Matt O'Reilly fermoymatt@gmail.com